+++ example/lib/main.dart
@@ -55,7 +55,7 @@ class MyApp extends StatelessWidget {
               GlobalMaterialLocalizations.delegate,
               GlobalWidgetsLocalizations.delegate,
               GlobalCupertinoLocalizations.delegate,
-              FlutterQuillLocalizations.delegate,
+              // FlutterQuillLocalizations.delegate,
             ],
             supportedLocales: FlutterQuillLocalizations.supportedLocales,
             routes: {